VPC test fixes for project import (#185)
* Minor fix for vpc project import

* black format

(cherry picked from commit f10b685ef1575fef2f4762c8e636b55a088359fc)
