+++ tests/tunnel_endpoints/http_ep_with_auth.py
@@ -6,7 +6,7 @@ from calm.dsl.builtins.models.metadata import Metadata
 from calm.dsl.runbooks import read_local_file
 from calm.dsl.runbooks import CalmEndpoint as Endpoint
 from calm.dsl.builtins.models.calm_ref import Ref
-from tests.tunnel_endpoints.helper import get_vpc_tunnel_using_account, get_vpc_project
+from tests.utils import get_vpc_tunnel_using_account, get_vpc_project
 
 AUTH_USERNAME = read_local_file(".tests/runbook_tests/auth_username")
 AUTH_PASSWORD = read_local_file(".tests/runbook_tests/auth_password")
